requirements.txt
streamlit>=1.32.0
plotly>=5.18.0
pandas>=2.0.0
python-dotenv>=1.0.0
langchain-google-genai>=1.0.0
google-generativeai>=0.4.0
langchain>=0.1.0
crewai>=0.28.0
